Implemented required abstract methods in Saver class to define its behavior
Implemented validateMarbleSize to ensure that exactly one marble is selected when using the Saver card.

Implemented validateMarbleColours to verify that the selected marble belongs to the active player.

Implemented act method to execute the Saver card's unique ability:
Moves one of the active player’s marbles from the track or base to a random empty safe zone cell.

Throws InvalidMarbleException if an invalid marble is selected.